+++ packages/kernel/src/declarations.d.ts
@@ -11,6 +11,6 @@ declare module "*.whl" {
 
 // Declarations for the `self` object in worker.ts some of whose properties are used to share values with the Python environment.
 interface Window {
-  __logCallback__: (msg: string) => void;
+  __logCallback__: (levelno: number, msg: string) => void;
   __streamlitFlagOptions__: Record<string, PyodideConvertiblePrimitive>;
 }